Implement Saving a Gig.

+++ GigHub/ViewModels/GigFormViewModel.cs
@@ -6,9 +6,13 @@ namespace GigHub.ViewModels
     public class GigFormViewModel
     {
         public string Venue { get; set; }
+        
         public string Date { get; set; }
+        
         public string Time { get; set; }
-        public int Genre { get; set; }
+        
+        public byte Genre { get; set; }
+        
         public IEnumerable<Genre> Genres { get; set; }
     }
 }
